requirements.txt
matplotlib==3.7.1
numpy==2.3.2
pandas==2.3.2
PyYAML==6.0
PyYAML==6.0.2
scikit_learn==1.7.1
seaborn==0.13.2
torch==2.6.0
xgboost==3.0.4
~orch==1.11.0
~orch==2.0.1
